Add more logical structuring of files for CallMapper. Update version and the respective badges. Update main README.md and CallMapper README.md
