+++ CallMapper/css/style.css
@@ -1,0 +1,192 @@
+
+html, body {
+    height: 100%;
+    margin: 0;
+    font-family: Verdana;
+}
+
+#cy {
+    position: relative; 
+    width: 100%;
+    height: 100%;
+}
+
+#label-hide-orphaned, #label-select-all{
+    font-weight: bold;
+}
+
+#invalid-endpoint-title{
+    display: flex;
+    justify-content: center; 
+    margin-bottom: -18px;
+    margin-top: 5px;
+    font-weight: bold;
+    font-size: 16px;
+    color: #5f5f5f;
+}
+
+#api-source-title{
+    display: flex;
+    justify-content: center; 
+    margin-bottom: -18px;
+    margin-top: 5px;
+    font-weight: bold;
+    font-size: 16px;
+    color: #000000;
+}
+
+#endpoint-type{
+    display: flex;
+    justify-content: center; 
+    margin-bottom: -30px;
+    margin-top: -5px;
+    font-size: 30px;
+    color: #000000;
+}
+
+#endpoint-value{
+    display: flex;
+    justify-content: center; 
+    color: rgb(0, 0, 0);
+    margin-bottom: -2px;
+    max-width: 100%; 
+    word-break: break-all; 
+}
+
+#deselect-button{
+    margin-top: 5px;
+    display: flex;
+    justify-content: center; 
+    color: white;
+    background-color: #cc00cc; 
+    padding: 10px 0px;
+    width: 100%;
+    border-radius: 5px;
+    border-width: 1px; 
+    border-color: #000000;
+    transition: background-color 0.2s ease, transform 0.1s ease;
+    font-weight: bold;
+}
+
+#deselect-button:hover{
+    background-color: #cc00ccab; 
+    transform: translateY(-1px); 
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); 
+}    
+
+#deselect-button:active {
+    background-color: #cc00cc62; 
+    transform: translateY(0); 
+}
+
+a {
+    margin-top: 5px;
+    margin-bottom: -10px;
+    display: flex;
+    justify-content: center; 
+    color: white;
+    background-color: #cc00cc; 
+    border: 1px solid #4b4b4b;
+    padding: 10px 0px;
+    width: 100%;
+    border-radius: 5px; 
+    text-decoration: none; 
+    transition: background-color 0.3s ease, transform 0.2s ease;
+}
+
+a:hover {
+    background-color: #cc00ccab; 
+    transform: translateY(-2px);
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1); 
+}
+
+a:active {
+    background-color: #cc00cc62; 
+    transform: translateY(0); 
+}
+
+#tooltip {
+    display: none;
+    position: absolute;
+    right: 0; 
+    border: 1px solid #ccc;
+    height: 100%;
+    width: 300px;
+    padding: 10px;
+    border-left: 1px solid #929292; 
+    box-shadow: 0px 4px 8px rgba(255, 110, 110, 0.1);
+    z-index: 100;
+    pointer-events: auto;
+
+    font-size: 14px;
+    color: #000000;
+    background-color: rgb(255, 255, 255);
+    word-wrap: break-word; 
+    overflow-wrap: break-word; 
+
+    white-space: normal;
+    overflow: hidden;
+}
+
+#checkbox-pane{
+    padding: 10px; 
+    border-right: 1px solid #ccc; 
+    width: 12%; 
+    overflow-y: auto;
+}
+
+#checkbox-pane input[type="checkbox"] {
+    display: none;  
+}
+
+#checkbox-header{
+    display: block;
+    width: calc(100% - 13px); 
+    margin: 40px 0;
+    text-align: center;
+}
+
+#checkbox-title{
+    font-size: 20px;
+    color: #000000;
+}
+
+#checkbox-application-name{
+    font-size: 15px;
+    color: #111111;
+    margin-top: -18px;
+}
+
+#checkbox-author{
+    font-size: 9px;
+    color: #a0a0a0;
+    margin-top: -10px;
+}
+
+#checkbox-pane label {
+    display: block;
+    width: calc(100% - 10px); 
+    box-sizing: border-box;
+    cursor: pointer;
+    padding: 8px 12px;
+    margin: 4px 0;
+    border: 1px solid #4b4b4b;
+    border-radius: 5px;
+    font-size: 14px;
+    color: #cc00cc;
+    background-color: white;
+    transition: background-color 0.3s ease, color 0.3s ease, box-shadow 0.3s ease;
+    text-align: left; 
+    white-space: normal;
+    word-wrap: break-word; 
+}
+
+#checkbox-pane input[type="checkbox"]:checked + label {
+    background-color: #cc00cc;
+    color: white;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+#checkbox-pane label:hover {
+    background-color: #cc00cc33;
+}
