+++ WhoYouCalling.csproj
@@ -6,7 +6,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <Platforms>AnyCPU;x64;x86</Platforms>
-    <FileVersion>1.4.3</FileVersion>
+    <FileVersion>1.5</FileVersion>
     <ApplicationManifest>WhoYouCalling\app.manifest</ApplicationManifest>
     <AssemblyName>wyc</AssemblyName>
   </PropertyGroup>
